Fix threshold-checking bug

+++ ddprofiler/app/src/main/java/ddprofiler/preanalysis/PreAnalyzer.java
@@ -231,7 +231,7 @@ public class PreAnalyzer implements PreAnalysis, IO {
                 // Check whether any spatial pattern matched > 50% of the values to determine the types.
                 // Assumption: no spatial and temporal patterns have >50% matches at the same time.
                 for (String spatialGranularity : spatialMatchCount.keySet()) {
-                    if ((long) spatialMatchCount.get(spatialGranularity) / validValuesCount > 0.5) {
+                    if ((double) spatialMatchCount.get(spatialGranularity) / validValuesCount > 0.5) {
                         attribute.setColumnSemanticType(AttributeSemanticType.SPATIAL);
                         attribute.getColumnSemanticTypeDetails().put("granularity", spatialGranularity);
                         break;
